+++ crates/matrix-sdk-crypto/src/machine/mod.rs
@@ -1720,9 +1720,9 @@ impl OlmMachine {
                     .iter()
                     .map(|(k, v)| (k.to_owned(), v.to_base64()))
                     .collect(),
+                session_id: Some(session.session_id().to_owned()),
             },
             verification_state,
-            session_id: Some(session.session_id().to_owned()),
         })
     }